--- a/pkg/agent/flowexporter/exporter/exporter.go
+++ b/pkg/agent/flowexporter/exporter/exporter.go
@@ -104,26 +104,20 @@ type flowExporter struct {
 	k8sClient           kubernetes.Interface
 	nodeRouteController *noderoute.Controller
 	isNetworkPolicyOnly bool
+	nodeName            string
 }
 
-func genObservationID() (uint32, error) {
-	name, err := env.GetNodeName()
-	if err != nil {
-		return 0, err
-	}
+func genObservationID(nodeName string) uint32 {
 	h := fnv.New32()
-	h.Write([]byte(name))
-	return h.Sum32(), nil
+	h.Write([]byte(nodeName))
+	return h.Sum32()
 }
 
-func prepareExporterInputArgs(collectorAddr, collectorProto string) (exporter.ExporterInput, error) {
+func prepareExporterInputArgs(collectorAddr, collectorProto, nodeName string) exporter.ExporterInput {
 	expInput := exporter.ExporterInput{}
-	var err error
 	// Exporting process requires domain observation ID.
-	expInput.ObservationDomainID, err = genObservationID()
-	if err != nil {
-		return expInput, err
-	}
+	expInput.ObservationDomainID = genObservationID(nodeName)
+
 	expInput.CollectorAddress = collectorAddr
 	if collectorProto == "tls" {
 		expInput.IsEncrypted = true
@@ -134,7 +128,7 @@ func prepareExporterInputArgs(collectorAddr, collectorProto string) (exporter.Ex
 	}
 	expInput.PathMTU = 0
 
-	return expInput, nil
+	return expInput
 }
 
 func NewFlowExporter(connStore *connections.ConntrackConnectionStore, records *flowrecords.FlowRecords, denyConnStore *connections.DenyConnectionStore,
@@ -146,10 +140,11 @@ func NewFlowExporter(connStore *connections.ConntrackConnectionStore, records *f
 	registry.LoadRegistry()
 
 	// Prepare input args for IPFIX exporting process.
-	expInput, err := prepareExporterInputArgs(collectorAddr, collectorProto)
+	nodeName, err := env.GetNodeName()
 	if err != nil {
 		return nil, err
 	}
+	expInput := prepareExporterInputArgs(collectorAddr, collectorProto, nodeName)
 
 	return &flowExporter{
 		conntrackConnStore:  connStore,
@@ -165,6 +160,7 @@ func NewFlowExporter(connStore *connections.ConntrackConnectionStore, records *f
 		k8sClient:           k8sClient,
 		nodeRouteController: nodeRouteController,
 		isNetworkPolicyOnly: isNetworkPolicyOnly,
+		nodeName:            nodeName,
 	}, nil
 }
 
@@ -334,7 +330,7 @@ func (exp *flowExporter) sendFlowRecords() error {
 				return err
 			}
 			exp.numDataSetsSent = exp.numDataSetsSent + 1
-			exp.denyConnStore.ResetConnStatsWithoutLock(conn)
+			exp.denyConnStore.ResetConnStatsWithoutLock(connKey)
 		}
 		if time.Since(conn.LastExportTime) >= exp.idleFlowTimeout {
 			if err := exp.addDenyConnToSet(conn, ipfixregistry.IdleTimeoutReason); err != nil {
@@ -414,8 +410,6 @@ func (exp *flowExporter) sendTemplateSet(isIPv6 bool) (int, error) {
 }
 
 func (exp *flowExporter) addRecordToSet(record flowexporter.FlowRecord) error {
-	nodeName, _ := env.GetNodeName()
-
 	// Iterate over all infoElements in the list
 	eL := exp.elementsListv4
 	if record.IsIPv6 {
@@ -488,7 +482,7 @@ func (exp *flowExporter) addRecordToSet(record flowexporter.FlowRecord) error {
 		case "sourceNodeName":
 			// Add nodeName for only local pods whose pod names are resolved.
 			if record.Conn.SourcePodName != "" {
-				ie.Value = nodeName
+				ie.Value = exp.nodeName
 			} else {
 				ie.Value = ""
 			}
@@ -499,7 +493,7 @@ func (exp *flowExporter) addRecordToSet(record flowexporter.FlowRecord) error {
 		case "destinationNodeName":
 			// Add nodeName for only local pods whose pod names are resolved.
 			if record.Conn.DestinationPodName != "" {
-				ie.Value = nodeName
+				ie.Value = exp.nodeName
 			} else {
 				ie.Value = ""
 			}
@@ -570,7 +564,6 @@ func (exp *flowExporter) addRecordToSet(record flowexporter.FlowRecord) error {
 }
 
 func (exp *flowExporter) addDenyConnToSet(conn *flowexporter.Connection, flowEndReason uint8) error {
-	nodeName, _ := env.GetNodeName()
 	exp.ipfixSet.ResetSet()
 
 	eL := exp.elementsListv4
@@ -622,7 +615,7 @@ func (exp *flowExporter) addDenyConnToSet(conn *flowexporter.Connection, flowEnd
 		case "sourceNodeName":
 			// Add nodeName for only local pods whose pod names are resolved.
 			if conn.SourcePodName != "" {
-				ie.Value = nodeName
+				ie.Value = exp.nodeName
 			} else {
 				ie.Value = ""
 			}
@@ -633,7 +626,7 @@ func (exp *flowExporter) addDenyConnToSet(conn *flowexporter.Connection, flowEnd
 		case "destinationNodeName":
 			// Add nodeName for only local pods whose pod names are resolved.
 			if conn.DestinationPodName != "" {
-				ie.Value = nodeName
+				ie.Value = exp.nodeName
 			} else {
 				ie.Value = ""
 			}