This commit adds performance benchmarking for Flow Exporter. It evaluates Export() function under different number of conntrack connections, dying connections, idle records, deny connections and idle deny connections. A local server will receive the records and count number. CPU and memory profile is collected and visualized using pprof.

Also from benchmarking, we discovered and removed redundant calls like `GetNodeName()`, which is called every time when exporting a record and `ResetConnStatsWithoutLock`, which unnecessarily calls `NewConnectionKey` each time.

With fixes and changes in go-ipfix [v0.5.3](https://github.com/vmware/go-ipfix/releases/tag/v0.5.3) and redundant calls removal in this PR, we had following improvements:
Before:
```
BenchmarkExportConntrackConns   	  50	  21747542 ns/op	 3708730 B/op	   80838 allocs/op
BenchmarkExportDenyConns      	          126	  14497807 ns/op	 2522699 B/op	   53345 allocs/op
BenchmarkPoll   	                 141	   8372127 ns/op	  813834 B/op	   52738 allocs/op
```
After
```
BenchmarkExportConntrackConns   	 126	  11444126 ns/op	  459816 B/op	   18332 allocs/op
BenchmarkExportDenyConns   	     223	   6391654 ns/op	  261669 B/op	   10637 allocs/op
BenchmarkPoll   	         130	   8280194 ns/op	  818076 B/op	   52814 allocs/op
```
Improvement
| Test   |  ns/op | B/op   | allocs/op |
| ------------------------- | --------- | --- | --- |
| BenchmarkExportConntrackConns  |  reduce by 47.4%	|  reduce by 87.6% | reduce by 77.3% | 
| BenchmarkExportDenyConns  |   reduce by 42.9%	|  reduce by 89.6%	|   reduce by 80.0% | 
| BenchmarkPoll  |  less than 1% diff	|  less than 1% diff	| less than 1% diff  |