@zyiou 2021-05-18
Sample metrics:
```
TOTAL_CONNECTIONS:   123551	
INIT_CONNECTIONS:   100000
UPDATE_CONNECTIONS/poll:   80000
NEW_CONNECTIONS/poll:   10000
DYING_CONNECTIONS/poll:   5000
POLL_INTERVAL(s):  1
TEST_DURATION(s):  10
MEMORY(M):  428 
```

Benchmark result:
```
go test -test.v -run=BenchmarkExport -test.benchmem -bench=. -memprofile memprofile.out -cpuprofile profile.out

goos: linux
goarch: amd64
pkg: antrea.io/antrea/pkg/agent/flowexporter/exporter
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkExport
BenchmarkExport-2   	       1	35014186189 ns/op	1322987648 B/op	27009979 allocs/op
```
```
go tool pprof memprofile.out
File: exporter.test
Type: alloc_space
Time: May 18, 2021 at 11:43pm (UTC)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) top
Showing nodes accounting for 1103.48MB, 86.50% of 1275.64MB total
Dropped 63 nodes (cum <= 6.38MB)
Showing top 10 nodes out of 50
      flat  flat%   sum%        cum   cum%
     480MB 37.63% 37.63%   783.52MB 61.42%  github.com/vmware/go-ipfix/pkg/entities.(*dataRecord).AddInfoElement
  116.50MB  9.13% 46.76%   210.52MB 16.50%  github.com/vmware/go-ipfix/pkg/entities.EncodeToIEDataType
  105.63MB  8.28% 55.04%   108.13MB  8.48%  antrea.io/antrea/pkg/agent/flowexporter/flowrecords.(*FlowRecords).AddOrUpdateFlowRecord
      93MB  7.29% 62.33%       93MB  7.29%  github.com/vmware/go-ipfix/pkg/entities.NewInfoElementWithValue (inline)
   66.01MB  5.17% 67.51%    66.01MB  5.17%  k8s.io/klog/v2.(*loggingT).header
   64.88MB  5.09% 72.59%   140.39MB 11.01%  antrea.io/antrea/pkg/agent/flowexporter/connections.(*connectionStore).addOrUpdateConn
   61.02MB  4.78% 77.38%    61.02MB  4.78%  bytes.makeSlice
   44.42MB  3.48% 80.86%    44.42MB  3.48%  antrea.io/antrea/pkg/agent/flowexporter/exporter.getUpdateConnections
   40.52MB  3.18% 84.03%    47.02MB  3.69%  antrea.io/antrea/pkg/agent/flowexporter/exporter.getConnections
   31.50MB  2.47% 86.50%    31.50MB  2.47%  net.IP.String
```