@zyiou 2021-05-25
We are having several performance improvements on go-ipfix side. Here is the performance difference for this [change](https://github.com/vmware/go-ipfix/pull/204)

Before change:
```
go test -test.v -run=BenchmarkExport -test.benchmem -bench=. -memprofile memprofile.out -cpuprofile profile.out

goos: linux
goarch: amd64
pkg: antrea.io/antrea/pkg/agent/flowexporter/exporter
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkExport
BenchmarkExport-2   	       1	20001730232 ns/op	1107956184 B/op	21170436 allocs/op
PASS
ok  	antrea.io/antrea/pkg/agent/flowexporter/exporter	20.131s
```
After change:
```
go test -test.v -run=BenchmarkExport -test.benchmem -bench=. -memprofile memprofile.out -cpuprofile profile.out
goos: linux
goarch: amd64
pkg: antrea.io/antrea/pkg/agent/flowexporter/exporter
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkExport
BenchmarkExport-2   	       1	20002124526 ns/op	801765208 B/op	13902598 allocs/op
PASS
ok  	antrea.io/antrea/pkg/agent/flowexporter/exporter	20.132s
```