@zyiou 2021-06-03
> In addition to the metrics, can we make a check if our memory utilization is linear with number of new connections added. Basically, the idea is to detect the memory leaks.

Sure. Will create a larger vm for testing.