@srikartati 2021-06-08
> > In addition to the metrics, can we make a check if our memory utilization is linear with number of new connections added. Basically, the idea is to detect the memory leaks.
> 
> Sure. Will create a larger vm for testing.

I just meant to change the initial number of connections through an input parameter (from a lower number to a higher number) and tracking memory utilization. Current number can be the maximum number.