@zyiou 2021-06-11
Test results with different number of new connections per poll:
| # new conns / poll  |   B/op   | allocs/op | max memory alloc(M) | 
| ------------------------- | ------- | --------- | ----- |
| 2000 | 1,872,791,432  | 38,548,557 | 431  | 
| 4000 | 2,525,452,400  | 52,035,182 | 576 | 
| 6000 | 3,076,454,200  | 63,339,188 | 726  | 
| 8000 | 3,730,222,696  | 77,111,198 | 899  | 
| 10000 | 4,432,394,112  | 90,846,552 | 1020  | 

Memory consumption is roughly linear as number of new connections increase.