@zyiou 2021-06-22
Updated testing result:
```
go test -test.v -run=BenchmarkExport -test.benchmem -bench=. -memprofile memprofile.out -cpuprofile profile.out
goos: linux
goarch: amd64
pkg: antrea.io/antrea/pkg/agent/flowexporter/exporter
cpu: Intel(R) Core(TM) i7-8750H CPU @ 2.20GHz
BenchmarkExport
    exporter_perf_test.go:61: 
        Summary:
        Number of conntrack connections: 20000
        Number of dying conntrack connections: 10000
        Number of deny connections: 20000
        Number of idle deny connections: 10000
        Total connections received: 37788
BenchmarkExport-2   	       1	1123878204 ns/op	135576880 B/op	 3458292 allocs/op
PASS
ok  	antrea.io/antrea/pkg/agent/flowexporter/exporter	1.298s
```
```
go tool pprof memprofile.out
File: exporter.test
Type: alloc_space
Time: Jun 22, 2021 at 2:01am (UTC)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) top
Showing nodes accounting for 103.32MB, 83.29% of 124.05MB total
Dropped 16 nodes (cum <= 0.62MB)
Showing top 10 nodes out of 68
      flat  flat%   sum%        cum   cum%
   22.50MB 18.14% 18.14%       27MB 21.77%  github.com/vmware/go-ipfix/pkg/entities.(*dataRecord).AddInfoElement
   15.76MB 12.70% 30.84%    15.76MB 12.70%  antrea.io/antrea/pkg/agent/flowexporter/flowrecords.(*FlowRecords).AddFlowRecordToMap
   15.50MB 12.50% 43.34%    15.50MB 12.50%  k8s.io/klog/v2.(*loggingT).header
   12.04MB  9.71% 53.05%    12.04MB  9.71%  antrea.io/antrea/pkg/agent/flowexporter/connections.(*connectionStore).AddConnToMap
    8.50MB  6.86% 59.91%    33.52MB 27.02%  antrea.io/antrea/pkg/agent/flowexporter/exporter.addConnsAndGetRecords
       8MB  6.45% 66.36%    11.50MB  9.27%  github.com/vmware/go-ipfix/pkg/exporter.(*ExportingProcess).createAndSendMsg
       7MB  5.65% 72.00%    15.29MB 12.33%  antrea.io/antrea/pkg/agent/flowexporter/exporter.addDenyConns
    5.50MB  4.43% 76.44%    35.50MB 28.62%  antrea.io/antrea/pkg/agent/flowexporter/exporter.(*flowExporter).addRecordToSet
    4.50MB  3.63% 80.06%     4.50MB  3.63%  github.com/vmware/go-ipfix/pkg/entities.EncodeToIEDataType
       4MB  3.22% 83.29%        4MB  3.22%  github.com/vmware/go-ipfix/pkg/entities.NewDataRecord (inline)
```

```
go tool pprof profile.out
File: exporter.test
Type: cpu
Time: Jun 22, 2021 at 2:01am (UTC)
Duration: 1.24s, Total samples = 1.11s (89.81%)
Entering interactive mode (type "help" for commands, "o" for options)
(pprof) top
Showing nodes accounting for 770ms, 69.37% of 1110ms total
Showing top 10 nodes out of 149
      flat  flat%   sum%        cum   cum%
     510ms 45.95% 45.95%      510ms 45.95%  syscall.Syscall
      60ms  5.41% 51.35%      100ms  9.01%  runtime.scanobject
      40ms  3.60% 54.95%       70ms  6.31%  runtime.mallocgc
      40ms  3.60% 58.56%       40ms  3.60%  syscall.RawSyscall
      20ms  1.80% 60.36%      160ms 14.41%  antrea.io/antrea/pkg/agent/flowexporter/exporter.(*flowExporter).addDenyConnToSet
      20ms  1.80% 62.16%       20ms  1.80%  cmpbody
      20ms  1.80% 63.96%       20ms  1.80%  runtime.heapBits.bits (inline)
      20ms  1.80% 65.77%       40ms  3.60%  runtime.makeslice
      20ms  1.80% 67.57%       20ms  1.80%  runtime.memmove
      20ms  1.80% 69.37%       20ms  1.80%  runtime.step
```
Benchmark result with different number of conntrack connections + deny connections input
| # conns | ns/op | B/op  | allocs/op |
| ------------------------- | ------- | --------- | ----|
| 20,000 + 20,000 | 1,123,878,204 | 135,576,880 |3,458,292|
| 50,000 + 50,000| 2,856,738,645 | 329,267,544 | 8,690,977 |
| 100000 + 100000|5,469,929,206| 656,425,016 | 17,396,982|